requirements.txt
flask
pandas
openpyxl
xlsxwriter
python-pptx
python-docx
pillow
scikit-learn
numpy